test(server): split coverage audit buckets
- Move durable coverage_gap_audit tests into permanent behavior modules beside the diagnostics, run, adapters, infra, and history owners
- Delete the temporary audit bucket files so the test names describe product behavior instead of audit provenance

Closes #2856
