+++ apps/server/tests/adapters/websocket/test_ws_hub_run_loop.py
@@ -1,0 +1,69 @@
+"""Behavior tests for the WebSocket hub run loop."""
+
+from __future__ import annotations
+
+import asyncio
+from unittest.mock import AsyncMock
+
+import pytest
+
+from vibesensor.adapters.websocket.hub import WebSocketHub
+
+
+class TestWSHubRunLoop:
+    """WebSocketHub.run() should tick, recover, and stop cleanly."""
+
+    @pytest.mark.asyncio
+    async def test_run_calls_on_tick_and_broadcasts(self) -> None:
+        hub = WebSocketHub()
+        ws = AsyncMock()
+        ws.send_text = AsyncMock()
+        await hub.add(ws, None)
+
+        tick_count = 0
+
+        def on_tick() -> None:
+            nonlocal tick_count
+            tick_count += 1
+
+        task = asyncio.create_task(
+            hub.run(hz=100, payload_builder=lambda _: {"ok": True}, on_tick=on_tick),
+        )
+        await asyncio.sleep(0.15)
+        task.cancel()
+        with pytest.raises(asyncio.CancelledError):
+            await task
+        assert tick_count >= 2
+        assert ws.send_text.await_count >= 2
+
+    @pytest.mark.asyncio
+    async def test_run_survives_broadcast_exception(self) -> None:
+        hub = WebSocketHub()
+        call_count = 0
+
+        def flaky_builder(_cid):
+            nonlocal call_count
+            call_count += 1
+            if call_count == 1:
+                raise RuntimeError("first tick boom")
+            return {"ok": True}
+
+        ws = AsyncMock()
+        ws.send_text = AsyncMock()
+        await hub.add(ws, None)
+
+        task = asyncio.create_task(hub.run(hz=50, payload_builder=flaky_builder))
+        await asyncio.sleep(0.15)
+        task.cancel()
+        with pytest.raises(asyncio.CancelledError):
+            await task
+        assert call_count >= 2
+
+    @pytest.mark.asyncio
+    async def test_run_hz_clamps_to_minimum_1(self) -> None:
+        hub = WebSocketHub()
+        task = asyncio.create_task(hub.run(hz=0, payload_builder=lambda _: {}))
+        await asyncio.sleep(0.05)
+        task.cancel()
+        with pytest.raises(asyncio.CancelledError):
+            await task
