+++ apps/server/tests/use_cases/diagnostics/test_order_confidence.py
@@ -1,0 +1,127 @@
+"""Direct behavior tests for order confidence scoring."""
+
+from __future__ import annotations
+
+from typing import Any
+
+import pytest
+
+from vibesensor.use_cases.diagnostics.orders.statistics import (
+    compute_order_confidence as _compute_order_confidence,
+)
+
+
+class TestComputeOrderConfidence:
+    """Direct unit tests for _compute_order_confidence."""
+
+    _DEFAULTS: dict[str, Any] = {
+        "effective_match_rate": 0.60,
+        "error_score": 0.80,
+        "corr_val": 0.50,
+        "snr_score": 0.60,
+        "absolute_strength_db": 20.0,
+        "localization_confidence": 0.70,
+        "weak_spatial_separation": False,
+        "dominance_ratio": 2.0,
+        "constant_speed": False,
+        "steady_speed": False,
+        "matched": 30,
+        "corroborating_locations": 2,
+        "phases_with_evidence": 2,
+        "is_diffuse_excitation": False,
+        "diffuse_penalty": 1.0,
+        "n_connected_locations": 3,
+        "no_wheel_sensors": False,
+        "path_compliance": 1.0,
+    }
+
+    @classmethod
+    def _call(cls, **overrides: Any) -> float:
+        return _compute_order_confidence(**{**cls._DEFAULTS, **overrides})
+
+    def test_baseline_returns_moderate_confidence(self) -> None:
+        conf = self._call()
+        assert 0.30 < conf < 0.90, f"Baseline defaults produced unexpected {conf}"
+
+    def test_output_clamped_low(self) -> None:
+        """All-zero inputs should clamp to the 0.08 floor."""
+        conf = self._call(
+            effective_match_rate=0.0,
+            error_score=0.0,
+            corr_val=0.0,
+            snr_score=0.0,
+            absolute_strength_db=0.0,
+            localization_confidence=0.0,
+            matched=0,
+            corroborating_locations=0,
+            phases_with_evidence=0,
+        )
+        assert conf == pytest.approx(0.08, abs=0.001)
+
+    def test_output_clamped_high(self) -> None:
+        """Perfect inputs should clamp to the 0.97 ceiling."""
+        conf = self._call(
+            effective_match_rate=1.0,
+            error_score=1.0,
+            corr_val=1.0,
+            snr_score=1.0,
+            absolute_strength_db=40.0,
+            localization_confidence=1.0,
+            matched=100,
+            corroborating_locations=4,
+            phases_with_evidence=4,
+        )
+        assert conf == pytest.approx(0.97, abs=0.001)
+
+    def test_negligible_strength_caps_at_045(self) -> None:
+        """absolute_strength_db below negligible threshold should cap confidence."""
+        conf = self._call(absolute_strength_db=5.0)
+        assert conf <= 0.45 + 0.001
+
+    @pytest.mark.parametrize(
+        ("normal_kw", "penalty_kw"),
+        [
+            pytest.param(
+                {"weak_spatial_separation": False},
+                {"weak_spatial_separation": True},
+                id="weak_spatial_separation",
+            ),
+            pytest.param(
+                {"constant_speed": False},
+                {"constant_speed": True},
+                id="constant_speed",
+            ),
+            pytest.param(
+                {"is_diffuse_excitation": False},
+                {"is_diffuse_excitation": True, "diffuse_penalty": 0.75},
+                id="diffuse_excitation",
+            ),
+            pytest.param(
+                {"n_connected_locations": 3},
+                {"n_connected_locations": 1},
+                id="single_sensor",
+            ),
+            pytest.param(
+                {"absolute_strength_db": 25.0},
+                {"absolute_strength_db": 12.0},
+                id="light_strength_band",
+            ),
+        ],
+    )
+    def test_penalty_reduces_confidence(
+        self,
+        normal_kw: dict[str, Any],
+        penalty_kw: dict[str, Any],
+    ) -> None:
+        assert self._call(**penalty_kw) < self._call(**normal_kw)
+
+    def test_path_compliance_shifts_weights(self) -> None:
+        """Higher path_compliance should shift weight from corr to match."""
+        stiff = self._call(path_compliance=1.0, corr_val=0.0, effective_match_rate=0.80)
+        compliant = self._call(path_compliance=1.5, corr_val=0.0, effective_match_rate=0.80)
+        assert compliant >= stiff - 0.02
+
+    def test_corroborating_locations_boost(self) -> None:
+        base = self._call(corroborating_locations=1)
+        boosted = self._call(corroborating_locations=3)
+        assert boosted > base, "3+ corroborating locations should boost confidence"
